We always expect the name of the created temporary file to be non-null; make this epectation explicit.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26023 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/arg/witnessexport/WitnessExporterTest.java
@@ -202,11 +202,14 @@ public class WitnessExporterTest {
               .suffix(pSuffix + compressedSuffix)
               .create()
               .toAbsolutePath();
-      String violationWitnessFileName = compressedFilePath.getFileName().toString();
-      String uncompressedWitnessFileName =
-          violationWitnessFileName.substring(
-              0, violationWitnessFileName.length() - compressedSuffix.length());
-      uncompressedFilePath = compressedFilePath.resolveSibling(uncompressedWitnessFileName);
+      Path compressedFileNamePath = compressedFilePath.getFileName();
+      if (compressedFilePath == null) {
+        throw new AssertionError("Files obtained from TempFile.builder().create() should always have a file name.");
+      }
+      String fileName = compressedFileNamePath.toString();
+      String uncompressedFileName =
+          fileName.substring(0, fileName.length() - compressedSuffix.length());
+      uncompressedFilePath = compressedFilePath.resolveSibling(uncompressedFileName);
       uncompressedFilePath.toFile().deleteOnExit();
     }
